requirements.txt
torch
einops
tqdm
coolname
pydantic
argdantic
wandb
omegaconf
hydra-core
huggingface_hub
numpy
matplotlib
datasets
PyYAML